+++ include/GafferScene/BranchCreator.h
@@ -89,6 +89,9 @@ class GAFFERSCENE_API BranchCreator : public FilteredSceneProcessor
 		Gaffer::StringPlug *destinationPlug();
 		const Gaffer::StringPlug *destinationPlug() const;
 
+		Gaffer::BoolPlug *copySourceAttributesPlug();
+		const Gaffer::BoolPlug *copySourceAttributesPlug() const;
+
 		void affects( const Gaffer::Plug *input, AffectedPlugsContainer &outputs ) const override;
 
 	protected :